hello_fastapi/follow_ithelp/pyproject.toml
[tool.poetry]
name = "follow-ithelp"
version = "0.1.0"
description = ""
authors = ["Eddy <eddy15201@gmail.com>"]
readme = "README.md"
packages = [{include = "follow_ithelp"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.106.0"
uvicorn = "^0.25.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.23"
cryptography = "^41.0.7"
aiomysql = "^0.2.0"
httpx = "^0.26.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
asyncpg = "^0.29.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
redis = "4.6.0"
redis-om = "^0.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.9"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.23.3"
pytest-sugar = "^0.9.7"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
